Add generic sha-2 (224, 256, 384, 512) implementations for convenience. (#56)
Add generic implementations of SHA 224, 256, 384, and 512.
This is a service the integrator of proxy dice services. Those
may not need the full suite of crypto primitives as provided by
the `crypto.h` interface but they still might need to digest DICE 
inputs.

- [x] Add doxygen documentation.
- [x] Double check the generated docs

Note: In a followup PR this implementation is used to implement
the n20_crypto_digest_context_t. The it is subjected to all the
crypto unit tests of the n20 crypto suite.