Implement crypto digest context in terms of software sha2 implementation (#61)
This pull request introduces the `nat20` cryptographic implementation to
the codebase.
This is a free standing implementation of the n20_crypto_digest_context
subset
of the crypto interface.

It provides an implementation of the digest function in terms of the
software
Implementation added in #56. The function hmac is implemented in terms
of this digest function and the hkdf family of function is implemented
in terms
of the hmac function.

- [x] The digest subset of the crypto test suite is run against the new
implementation.